Move python/tlm_adjoint_fenics/ to tlm_adjoint/fenics/
