+++ docs/manual.tex
@@ -121,7 +121,7 @@ A forward model constrained derivative of $J$ with respect to the discrete
 initial condition $\psi^0$ can computed via
 \begin{lstlisting}
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 stop_manager()
 
 N = 100
@@ -256,7 +256,7 @@ derivative of the functional $J$ with respect to the initial condition
 $\psi^0$, in a direction $\zeta \in V$, can be computed via
 \begin{lstlisting}
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 stop_manager()
 
 N = 100
@@ -367,7 +367,7 @@ For further details see section \ref{sect:higher_order}.
 The FEniCS backend may be used via, e.g.
 \begin{lstlisting}
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 \end{lstlisting}
 
 The Firedrake backend may be used via, e.g.
@@ -415,7 +415,7 @@ solve(inner(test, trial) * dx == inner(test, F * F) * dx,
 A forward model constrained derivative can be computed via, for example
 \begin{lstlisting}
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 
 mesh = UnitIntervalMesh(10)
 space = FunctionSpace(mesh, "Lagrange", 1)
@@ -2485,7 +2485,7 @@ example, the diffusion equation example described in section
 \ref{sect:diffusion} may be implemented via
 \begin{lstlisting}
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 from tlm_adjoint.timestepping import *
 stop_manager()
