+++ examples/fenics/basal_sliding/basal_fp.py
@@ -19,7 +19,7 @@
 # along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
 
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 from tlm_adjoint.hessian_optimization import *
 
 # import h5py
