+++ examples/fenics/manual/override_adjoint.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 
 mesh = UnitIntervalMesh(10)
 space = FunctionSpace(mesh, "Lagrange", 1)
