+++ examples/fenics/poisson/poisson.py
@@ -19,7 +19,7 @@
 # along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
 
 from fenics import *
-from tlm_adjoint_fenics import *
+from tlm_adjoint.fenics import *
 
 import numpy as np
 # import petsc4py.PETSc as PETSc
