+++ tests/fenics/test_manager.py
@@ -19,8 +19,8 @@
 # along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
 
 from fenics import *
-from tlm_adjoint_fenics import *
-from tlm_adjoint_fenics import manager as _manager
+from tlm_adjoint.fenics import *
+from tlm_adjoint.fenics import manager as _manager
 from tlm_adjoint.alias import WeakAlias
 
 from test_base import *
